unify results checks for various frameworks into shared getVisualResults utility
