--- visual-js/.changeset/swift-rings-sip.md
@@ -1,6 +1,0 @@
----
-"@saucelabs/visual-playwright": minor
-"@saucelabs/visual-storybook": patch
----
-
-Add ability to assert on visual results in Playwright
