+++ visual-js/visual-playwright/package.json
@@ -33,9 +33,11 @@
     "test": "jest --collect-coverage"
   },
   "dependencies": {
-    "@playwright/test": "^1.42.1",
     "@saucelabs/visual": "^0.9.0",
     "exponential-backoff": "^3.1.1"
+  },
+  "peerDependencies": {
+    "@playwright/test": "^1.42.1"
   },
   "tsup": {
     "entry": [
@@ -54,6 +56,7 @@
   },
   "devDependencies": {
     "@jest/globals": "^28.0.0 || ^29.0.0",
+    "@playwright/test": "^1.48.0",
     "@storybook/types": "^8.0.2",
     "@tsconfig/node18": "^2.0.0",
     "@types/node": "^18.13.0",
